Use proper key for hiding host_browser from launcher
I was using 'Hidden=true' which is incorrect usage of that so I've
changed it to use 'NoDisplay=true' which will also prevent it from being
shown in the launcher. I also added 'OnlyShowIn=Never' which will then
also prevent it from getting sent over to Chrome simply for mime type
registration (when that's added later).

Change-Id: I5207e886db9be9a895a7e899cadf10b2ede354bd
Reviewed-on: https://chromium-review.googlesource.com/1022112
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>