cros-sommelier-config: Add symlinks for swrast_dri.so
Xwayland hardcodes the search location for swrast_dri.so to be
/usr/lib64/dri/swrast_dri.so since that directory is where all the dri
drivers for chrome os are installed.

Have the cros-sommelier-config package install a symlink in this
location that points back to the /opt/google location where the library
is actually installed.

This should be reverted once we have a proper fix for mesa to load the
library from the right location.

BUG=chromium:835959
TEST=`glxinfo` works and `xdpyinfo | grep GLX` shows output inside the
     container

Change-Id: I0cae0249a01be37468e75e0dce46395d5036633a
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1024878