cros-tast-tests: Add StartupWMClass properties and icons for demo apps
This adds icons which are just solid color 32x32 pngs with the fill
color matching the painting color we are using for the tests. These
won't be evaluated graphically but will be checked for existence in the
container as well as being fetched properly by Chrome. It also sets the
StartupWMClass so Chrome can link it to the window properly.

BUG=chromium:849438
TEST=Verified apps link to shelf icons on eve

Change-Id: I0800e4ef3001586fe2f0aa9854a3fec6a0473eee
Reviewed-on: https://chromium-review.googlesource.com/1228268
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>