cros-garcon: Set XCURSOR_THEME for flatpak apps
This fixes a bug where flatpak apps would have tiny cursors because they
weren't using the cursor theme. Flatpak is supposed to default to the
Adwaita theme...but is not using it for the cursor for some reason.
Setting this in the environment fixes the problem.

Note that this fix will not always work until crbug.com/894276 is also
fixed.

BUG=chromium:889747
TEST=Manually verified on eve

Change-Id: I4a77ef76f76a59507cd2ad00b21e7f2b09f52663
Reviewed-on: https://chromium-review.googlesource.com/1275085
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: David Reveman <reveman@chromium.org>