Add synaptic auth and Xauthority creation
This is part of the fix for synaptic to work in the container. It adds
the polkit permission required for it to run and it also creates the
.Xauthority file when sommelier starts up so it can use the X server
when it has elevated its privileges to root. There's a corresponding
change in sommelier where it'll pass the .Xauthority file to Xwayland so
it will use it as well.

BUG=chromium:876846
TEST=Manually verified with synaptic on both X servers (0 and 1)

Change-Id: I26e426332d7ce3993fd93da0f1d3adb42072cb93
Reviewed-on: https://chromium-review.googlesource.com/1279069
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>