cros-sommelier: Don't erase .Xauthority file during startup
When I originally did this change, there was a discussion around using
'touch ~/.Xauthority' vs. using '>~/.Xauthority' and we determined the
latter was better because than if there was any corruption in the
.Xauthority file it would clean that up.  However, I've been seeing some
of the tests for the x11_demo app fail and I reproduced it locally and
noticed that the .Xauthority file was only 52 bytes instead of 104...and
it was missing the entry for the 0 display server. Since we are starting
up 2 instances of sommelier-x now (for low and high density) each of
them erases the .Xauthority file and then writes their entry. So we have
a clear race condition here.

It seems like it's a better option to assume that xauth itself will
clear up any issues with the .Xauthority file rather than having this
known race in there which will frequently cause issues.

BUG=chromium:903876
TEST=Verified .Xauthority file is intact multiple times

Change-Id: Id05db53b8a53bb79e6b15c7cbc5103abd43ccb68
Reviewed-on: https://chromium-review.googlesource.com/1336407
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: David Reveman <reveman@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>