cros-sommelier-config: Make DPI buckets only on low-DPI mode
Given that low-dpi mode is for supporting applications that need DPI
correction, it seems odd to force high-dpi applications to have DPI
buckets.

We are adding this configuration option to the low density overrides, so
that we can eventually make the default behaviour of sommelier "do not
bucket the DPI", and instead make it do that only in low-dpi mode (since
doing so is the kind of thing you do to support apps with poor scaling.

BUG=chromium:988325
TEST=manually verified it compiles

Change-Id: I22b1923ebcb142ac4c390883c4e39f9c3ff58d15
Reviewed-on: https://chromium-review.googlesource.com/1772958
Tested-by: Nic Hollingum <hollingum@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Ready: Nic Hollingum <hollingum@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>