termina_build_image: simplify rootfs copy
There was a loop over the files in the rootfs, but the inside of the
loop invoked rsync for the whole rootfs (not the individual files), so
the whole rootfs would get copied many times.

Drop the loop and just let rsync handle recursing over the files and
directories.

BUG=None
TEST=termina_build_image.py chromiumos_test_image.bin /tmp/termina

Change-Id: I43eb368dff442a0df5acede097c8b000c4a9c1fb
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1778579
Tested-by: kokoro <noreply+kokoro@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>