termina_build_image: split VM tools into extra disk
Allow reuse of the tools in /opt/google/cros-containers by moving them
to a separate disk image that can be used independently from the usual
Termina rootfs (vm_rootfs.img).

This works because all of the executables in this directory are
installed using lddtree to distribute all of the required libraries so
they can run without installing additional dependencies.

Note that this will require a corresponding concierge change to add the
extra disk image as a crosvm parameter and mount it in the expected
location inside the VM.

BUG=chromium:998033
TEST=termina_build_image.py chromiumos_test_image.bin /tmp/termina
TEST=Boot Crostini with modified rootfs + tools disk

Change-Id: I899072ab60874b3407c52b28e32878e315afb88c
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1778580
Tested-by: kokoro <noreply+kokoro@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>