termina_build_image: size image based on source
Rather than hard-coding a maximum 400 MB image size, which will need to
be adjusted later, use 120% the size of the source directory as a
starting point.  The image will still be shrunk to the minimum by
resize2fs before it is finalized.

This is equivalent to https://crrev.com/c/1544864 from the bash version
of this script, except that this uses 120% as the heuristic instead of
110%; 110% was failing for smaller images (resize2fs refuses to shrink
them and exits with an error).

BUG=chromium:999698
TEST=termina_build_image.py chromiumos_test_image.bin /tmp/termina

Change-Id: I6260f60686a4f122933c324f3ce95d062d18d7f5
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1779501
Tested-by: kokoro <noreply+kokoro@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>