Make libdrm and mesa build sequentially in the same shard
The partially-completed crrev.com/c/1925907 sharded the mesa builds
incorrectly, meaning mesa would be built against a potentially
out-of-date libdrm. The correct behaviour is to build libdrm in the same
shard as mesa.

This CL makes that fix, as well as several smaller fixes suggested in
the other CL.

The job configuration is changed in cl/282267687.

BUG=chromium:1018519
TEST=(if anyone knows how to test this i'm all ears)

Change-Id: Ic051e8920d11a3903cbdd40226af974a03f26945
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/1933610
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>