Share fonts from ChromiumOS to the GuestOS
We want the GuestOS to have access to the same fonts as ChromeOS,
after the font directory has been mounted to /mnt/chromeos/fonts, the
search directories for available fonts needs to see the above directory.

This is done through the installation of a Debian package which will
move a config file to /etc/fonts/conf.d that will specify a new path.

BUG=chromium:903631
TEST=create deb package using bazel, sudo apt-install and check
fc-list available fonts

Change-Id: Ic5a3c60350a040f6739b5e6c919a10e61652de40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/1980189
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Matthew Chen <matterchen@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>