container-guest-tools: Define a CI job to act as the entrypoint.
The kokoro CI is currently disabled (cl/286950668) however since the fix
for b/146575733 is not forthcoming, we will instead implement the
workaround.

We create an empty no-op job script and configuration file to be the new
head of the build. Once this lands, we can add a step to the kokoro
configuration in piper to use the "ci" job as the head.

BUG=b:146575733
TEST=none

Change-Id: Ieadf63ace38d5b3459d2ca5f4fa51439efc839bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/2006628
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>