kokoro: Allow presubmits to use the real build path
Currently the kokoro presubmits use a cut down version of the
container build that skips building the cros versions of debian
packages. This has caused the presubmits to break since the cros
version of apitrace started requiring a later version of libwaffle
then is supplied in the debian repos.

This change, along with a corresponding change to the internal
configs, will make the presubmit flow the same as the normal build,
except that it will skip uploading the results to GCS and won't sign
anything.

The presubmits will cover all combinations of OS versions and
archetectures both to ensure full coverage and because this step takes
<1/3 of the total build time, so reducing it has fairly marginal
gains.

BUG=none
TEST=untestable!

Change-Id: I098d0396effd75be3830d866431c1630627a976e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/2176817
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>