termina: Upload debug symbols in the termina uprev scripts
It turns out that the change I made to the {tatl,tael}_full builders
to upload debug symbols only uploads them to the staging crash server,
not the prod crash server. Since we need to be able to upload symbols
for manually triggered builds anyway, add uploading to the prod crash
server to the uprev scripts.

AFAICT the API key is used only for authentication, the crash server
doesn't e.g. categorise symbol files by who uploaded them, so I
enabled the API on the crosvm-packages cloud project and generated a
key for us to use.

BUG=none
TEST=Uploaded new build and check that symbols are present on the prod server

Change-Id: I97c4aa741b2e40a5e35d9a3a0875cf4d40f5ed91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/2262473
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>