container-guest-tools: Deal with gsutil
gsutil cp -r has semantics that depend on whether the target directory
already exists or not. If it does, the source directory will be
created under the target, otherwise the *contents* of the source
directory are copied directly under the (newly created) target. Make
sure we always target a directory that (shouldn't) exist by including
the build number in the target path.

BUG=none
TEST=Ran script

Disallow-Recycled-Builds: test-failures
Change-Id: Idd362a72e8d6533142613d8f08b77430f63fbd89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/2636034
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>