termina_build_image: vm_tools.img as cros-vm-tools
This set the ext2 label for the filesystem to "cros-vm-tools" which
makes it easy to find the fs in the guest regardless of the underlying
virtual hardware it is exposed on by just mounting
`/dev/disk/by-label/cros-vm-tools`.

This is useful for custom termina-like OSes since the termina approach
of having maitred told where to mount its tools image from is rather
unweildy when trying to fit the the whole thing into an init setup like
systemd. Instead such guests can just create automount files that
automatically find the guest tooling image by label and mount it to
/opt/google/cros-containers.

BUG=b:189121169
TEST=manually tested creating an image

Change-Id: I4e33c17767262fe3257ff6579e5d87881397425b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/2912037
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>