cros-apt-config: Export only the specific key we trust
This has two benefits: first, the google public keyring includes a
legacy 1024 bit key from 2007, and we have no reason to give this full
trust. Second, it reduces the burden on SWEs updating the keyring here
to validate it: gpg will only export the single master key
we trust, and sub-keys validly signed by it, so a hypothetical actor
who could modify the file as it is downloaded, but not access the
master key, would not be able to insert their key here, and so it
would not end up trusted by clients.

BUG=none
TEST=none

Change-Id: Iaba879bd1f724d262a92eaf5302d12ab7aba1f91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/3458703
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: James Ye <jamesye@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>