cros-garcon: Deprecate cros-garcon-override.conf
Since cros-garcon-override.conf is registered as a conffile, in the case
that a user has locally modified this file, their changes will not be
overwritten automatically with updated versions.

Customisations are not intended to be made to this file (see
crbug.com/927303). We deprecate this file by moving its contents to
cros-garcon.service, which WILL be overwritten on updates. In its place
we put a comment that points to our FAQ on custom environment variables.

Going forward, updates to garcon environment variables should be made
directly to cros-garcon.service. Note that if a modified
cros-garcon-override.conf is present, it will still take precedence if
a variable is present in both files.

BUG=b:227683168
TEST=Installed cros-garcon=0.26 via PackageKit, verified no change in
     environment variables and local modifications to
     cros-garcon-override.conf are preserved.

Change-Id: Ied471515d862e46c3e09c1128b799eddc6179453
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/3562598
Reviewed-by: Timothy Loh <timloh@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: James Ye <jamesye@google.com>