cros-garcon: Set XDG_CURRENT_DESKTOP to enable Firefox CSD
When running on X11, Firefox checks the value of XDG_CURRENT_DESKTOP to
determine whether Client-Side Decorations (tabs in titlebar) can be
used. We didn't set this, so they couldn't be used.

We use the value X-Generic to avoid changing the behaviour of other
programs that use this variable, like xdg-utils.

BUG=b:215259893
TEST="Title Bar" checkbox in "Customize Firefox" is visible, can be
     disabled and removes the system titlebar.

Change-Id: I1bf523e44f06a2af909a6dde7ea2345fdefb8742
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/3565265
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: James Ye <jamesye@google.com>
Commit-Queue: James Ye <jamesye@google.com>