cros-host-fonts: Don't unmount fonts on removal
A package upgrade is essentially a removal followed by an install, which
results in the systemd mount unit being stopped. Since this mount will
be created by LXD, there will be no way to re-create it within the
container. To maintain compatibility, the unit file is kept and started
on install, but disabled at boot.

If the old /mnt/chromeos/fonts share does not exist, and this package is
upgraded from the previous version, fonts may be broken for the current
boot of the container, so the old share should be kept until the
majority of users have upgraded.

BUG=b:231252066
TEST=bazel build //cros-host-fonts:deb
TEST=on DUT: dpkg -i cros_host_fonts_0.4_all.deb

Change-Id: Ifa758b86673328bd4fe40d312cad0234d037dea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/3849742
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: James Ye <jamesye@google.com>
Reviewed-by: David Munro <davidmunro@google.com>