cros-sommelier: Use container_token as VM_IDENTIFIER env var
This CL updates sommelier's start script to use
/dev/.container_token as the vm_identifier via the environment variable
VM_IDENTIFIER. This allows sommelier to add the token to the app window
IDs and is used to identify which guest the app window belongs to later
on the Chrome side, as described in go/guestos-shelf-icons. Note that
in the case where /dev/.container_token is not available or empty,
"termina" will continue to be used as the default vm_identifier.

Note that the container_token is set via environment variable rather
than the --vm-identifier cmdline arg because environment variables will
persist when sommelier forks, whereas persisting ctx.vm_id requires more
work.

BUG=b:244651040
TEST=manual testing as described in crrev.com/c/4173316

Change-Id: I1f4c50836cbaf7d760c45d3f0f31b2eee3973e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/4145286
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Sophia Lin <sophialin@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>