cros-guest-tools: Move cros-im from Recommends to bullseye depends
New Recommended packages are not installed when the base package is
updated, only when installing it for the first time. This CL moves
cros-im to Depends so users will be able to use it from existing
containers without having to manually force install it.

cros-im is only built for bullseye, so cros-guest-tools for buster will
break if it depends on the cros-im package. To resolve this, this CL
splits cros-guest-tools so a separate version is built for buster and
bullseye, similar to cros-apt-config.

BUG=b:255702607
TEST=kokoro, bazel build //cros-debs:deb and installing cros-guest-tools
on DUT in both bullseye and buster containers

Change-Id: I3e8bb9eefc7e886f7776e02a50807c5d0e8bf06c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/containers/cros-container-guest-tools/+/4178193
Reviewed-by: James Ye <jamesye@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Sophia Lin <sophialin@google.com>